// web/src/openapi/models/AppLatestInfoDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * tracing-lv-server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AppLatestInfoDto
 */
export interface AppLatestInfoDto {
    /**
     * 
     * @type {Date}
     * @memberof AppLatestInfoDto
     */
    buildTime: Date;
    /**
     * 
     * @type {string}
     * @memberof AppLatestInfoDto
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof AppLatestInfoDto
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof AppLatestInfoDto
     */
    nodeCount: number;
    /**
     * 
     * @type {number}
     * @memberof AppLatestInfoDto
     */
    onlineNodeCount: number;
}

/**
 * Check if a given object implements the AppLatestInfoDto interface.
 */
export function instanceOfAppLatestInfoDto(value: object): value is AppLatestInfoDto {
    if (!('buildTime' in value) || value['buildTime'] === undefined) return false;
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('nodeCount' in value) || value['nodeCount'] === undefined) return false;
    if (!('onlineNodeCount' in value) || value['onlineNodeCount'] === undefined) return false;
    return true;
}

export function AppLatestInfoDtoFromJSON(json: any): AppLatestInfoDto {
    return AppLatestInfoDtoFromJSONTyped(json, false);
}

export function AppLatestInfoDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): AppLatestInfoDto {
    if (json == null) {
        return json;
    }
    return {
        
        'buildTime': (new Date(json['build_time'])),
        'id': json['id'],
        'name': json['name'],
        'nodeCount': json['node_count'],
        'onlineNodeCount': json['online_node_count'],
    };
}

  export function AppLatestInfoDtoToJSON(json: any): AppLatestInfoDto {
      return AppLatestInfoDtoToJSONTyped(json, false);
  }

  export function AppLatestInfoDtoToJSONTyped(value?: AppLatestInfoDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'build_time': ((value['buildTime']).toISOString()),
        'id': value['id'],
        'name': value['name'],
        'node_count': value['nodeCount'],
        'online_node_count': value['onlineNodeCount'],
    };
}